+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_bundle/config/model/CommandBundleConfigModel.java
@@ -16,48 +16,50 @@ public class CommandBundleConfigModel {
         """)
     public List<BundleCommandNode> entries = new ArrayList<>() {
         {
-            /* level 4 commands */
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "my-command test-the-command-with-optional-arg <int int-arg-name> [str str-arg-name this is the default value]", List.of("say hello %player:name%", "say int is $int-arg-name", "say str is $str-arg-name")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "my-command test-the-command-with-literal-arg first-literal second-literal <str str-arg-name>", List.of("say hello %player:name%", "say str is $str-arg-name")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "my-command test-the-command-with-optional-arg-and-literal-arg <int int-arg-name> first-literal [str str-arg-name the default value can contains placeholder %player:name% in %world:name%]", List.of("say hello %player:name%", "say int is $int-arg-name", "say str is $str-arg-name")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "my-command test-the-command-with-a-greedy-string <int int-arg-name> first-literal [greedy-string greedy-string-arg-name this is the default value]", List.of("say hello %player:name%", "say int is $int-arg-name", "say str is $greedy-string-arg-name")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "give-apple-to-random-player", List.of("give %fuji:random_player% minecraft:apple %fuji:random 16 32%")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "shoot <entity-type entity-type-arg-name>", List.of("execute as %player:name% run summon $entity-type-arg-name ~ ~1 ~ {ExplosionPower:4,Motion:[3.0,0.0,0.0]}")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "strike", List.of("execute as %player:name% at @s run summon lightning_bolt ^ ^ ^10")));
+            /* Register `/my-command` commands. */
+            this.add(new BundleCommandNode("This command is used to test the `optional argument`.", new CommandRequirementDescriptor(4, null), "my-command a-command-with-optional-arg <int int-arg-name> [str str-arg-name this is the default value]", List.of("say hello %player:name%", "say int is $int-arg-name", "say str is $str-arg-name")));
+            this.add(new BundleCommandNode("This command is used to test the `literal argument` and `required argument`.", new CommandRequirementDescriptor(4, null), "my-command a-command-with-required-arg first-literal second-literal <str str-arg-name>", List.of("say hello %player:name%", "say str is $str-arg-name")));
+            this.add(new BundleCommandNode("This command is used to test the `greedy string` argument type.", new CommandRequirementDescriptor(4, null), "my-command a-command-with-a-greedy-string <int int-arg-name> first-literal [greedy-string greedy-string-arg-name this is the default value]", List.of("say hello %player:name%", "say int is $int-arg-name", "say str is $greedy-string-arg-name")));
+            this.add(new BundleCommandNode("This command is used to test the `literal argument`, `required argument` and `optional argument`.", new CommandRequirementDescriptor(4, null), "my-command a-command-with-all-types-of-args <int int-arg-name> first-literal [str str-arg-name the default value can contains placeholder %player:name% in %world:name%]", List.of("say hello %player:name%", "say int is $int-arg-name", "say str is $str-arg-name")));
 
-            /* Game mode change. */
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "gm <gamemode gamemode-arg>", List.of("run as player %player:name% gamemode $gamemode-arg")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "gmc", List.of("run as player %player:name% gamemode creative")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "gms", List.of("run as player %player:name% gamemode survival")));
+            /* Register gamemode switcher commands. */
+            this.add(new BundleCommandNode("This command is an alias for `/gamemode`", new CommandRequirementDescriptor(4, null), "gm <gamemode gamemode-arg>", List.of("run as player %player:name% gamemode $gamemode-arg")));
+            this.add(new BundleCommandNode("This command is an alias for `/gamemode creative`", new CommandRequirementDescriptor(4, null), "gmc", List.of("run as player %player:name% gamemode creative")));
+            this.add(new BundleCommandNode("This command is an alias for `/gamemode survival`", new CommandRequirementDescriptor(4, null), "gms", List.of("run as player %player:name% gamemode survival")));
+            this.add(new BundleCommandNode("This command is an alias for `/gamemode spectator`", new CommandRequirementDescriptor(4, null), "gmsp", List.of("run as player %player:name% gamemode spectator")));
 
-            /* Weather change. */
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "sun", List.of("weather clear")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "rain", List.of("weather rain")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "thunder", List.of("weather thunder")));
+            /* Register weather changer commands. */
+            this.add(new BundleCommandNode("This command is an alias for `/weather clear`", new CommandRequirementDescriptor(4, null), "sun", List.of("run as player %player:name% weather clear")));
+            this.add(new BundleCommandNode("This command is an alias for `/weather rain`", new CommandRequirementDescriptor(4, null), "rain", List.of("run as player %player:name% weather rain")));
+            this.add(new BundleCommandNode("This command is an alias for `/weather thunder`", new CommandRequirementDescriptor(4, null), "thunder", List.of("run as player %player:name% weather thunder")));
 
-            /* Time change. */
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "day", List.of("time set day")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "night", List.of("time set night")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "midnight", List.of("time set midnight")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "moon", List.of("time set moon")));
+            /* Register time changer commands. */
+            this.add(new BundleCommandNode("This command is an alias for `/time set day`", new CommandRequirementDescriptor(4, null), "day", List.of("run as player %player:name% time set day")));
+            this.add(new BundleCommandNode("This command is an alias for `/time set night`", new CommandRequirementDescriptor(4, null), "night", List.of("run as player %player:name% time set night")));
+            this.add(new BundleCommandNode("This command is an alias for `/time set midnight`", new CommandRequirementDescriptor(4, null), "midnight", List.of("run as player %player:name% time set midnight")));
+            this.add(new BundleCommandNode("This command is an alias for `/time set noon`", new CommandRequirementDescriptor(4, null), "noon", List.of("run as player %player:name% time set noon")));
 
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "unbreakable", List.of("run as player %player:name% enchant %player:name% minecraft:unbreaking")));
+            /* Register NBT commands. */
+            this.add(new BundleCommandNode("This command is an alias for `/data get entity`", new CommandRequirementDescriptor(4, null), "nbt entity <entity target>", List.of("run as fake-op %player:name% data get entity $target")));
+            this.add(new BundleCommandNode("This command is an alias for `/data get block`", new CommandRequirementDescriptor(4, null), "nbt block <block-pos target>", List.of("run as fake-op %player:name% data get block $target")));
+            this.add(new BundleCommandNode("This command is an alias for `/data get entity @s SelectedItem`", new CommandRequirementDescriptor(4, null), "nbt item", List.of("run as fake-op %player:name% data get entity %player:name% SelectedItem")));
 
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "move-speed set <double double-arg>", List.of("run as player %player:name% attribute %player:name% minecraft:generic.movement_speed base set $double-arg")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(4, null), "move-speed reset", List.of("run as player %player:name% attribute %player:name% minecraft:generic.movement_speed base set 0.10000000149011612")));
-
-
-            /* level 0 commands */
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(0, null), "introduce-me", List.of("run as fake-op %player:name% me i am %player:name%")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(0, null), "block-info <blockpos blockpos-arg-name>", List.of("run as fake-op %player:name% data get block $blockpos-arg-name")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(0, null), "entity-info <entity entity-arg-name>", List.of("run as fake-op %player:name% data get entity $entity-arg-name")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(0, null), "dice", List.of("say %player:name% just roll out %fuji:random 1 6% points.")));
-
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(0, null), "plugins", List.of("send-message %player:name% Server Plugins (0): ")));
-            this.add(new BundleCommandNode(new CommandRequirementDescriptor(0, null), "icanhasbukkit",
+            /* Register easter eggs commands. */
+            this.add(new BundleCommandNode("This command is an easter egg for `/plugins` in Bukkit.", new CommandRequirementDescriptor(0, null), "plugins", List.of("send-message %player:name% Server Plugins (0): ")));
+            this.add(new BundleCommandNode("This command is an easter egg for `/icanhasbukkit` in Bukkit.", new CommandRequirementDescriptor(0, null), "icanhasbukkit",
                 List.of("send-message %player:name% <i>Checking version, please wait..."
                 , "delay 2 send-message %player:name% This server is running Bukkit version (MC: %server:version%)"
                 , "delay 3 send-message %player:name% <green>You are running the latest version")));
+
+            /* Register utility commands. */
+            this.add(new BundleCommandNode("This command summons an entity with specified entity type, with an initial motion.", new CommandRequirementDescriptor(4, null), "shoot <entity-type entity-type-arg-name>", List.of("execute as %player:name% run summon $entity-type-arg-name ~ ~1 ~ {ExplosionPower:4,Motion:[3.0,0.0,0.0]}")));
+            this.add(new BundleCommandNode("This command summons a lightning_bolt entity.", new CommandRequirementDescriptor(4, null), "strike", List.of("execute as %player:name% at @s run summon lightning_bolt ^ ^ ^32")));
+            this.add(new BundleCommandNode("This command applies the `unbreakable` enchantment for the item in hand.", new CommandRequirementDescriptor(4, null), "unbreakable", List.of("run as player %player:name% enchant %player:name% minecraft:unbreaking")));
+            this.add(new BundleCommandNode("This command modifies your movement_speed attribute.", new CommandRequirementDescriptor(4, null), "move-speed set <double double-arg>", List.of("run as player %player:name% attribute %player:name% minecraft:generic.movement_speed base set $double-arg")));
+            this.add(new BundleCommandNode("This command modifies your movement_speed attribute.", new CommandRequirementDescriptor(4, null), "move-speed reset", List.of("run as player %player:name% attribute %player:name% minecraft:generic.movement_speed base set 0.10000000149011612")));
+            this.add(new BundleCommandNode("This command will introduce yourself to others.", new CommandRequirementDescriptor(4, null), "introduce-me", List.of("run as fake-op %player:name% me i am %player:name%")));
+            this.add(new BundleCommandNode("This command will roll a random dice.", new CommandRequirementDescriptor(4, null), "dice", List.of("say %player:name% just roll out %fuji:random 1 6% points.")));
+
         }
     };